Allow more bytes to be read from registers
The address counter does not increment while reading from the FIFO data register. It can contain up to 512 bytes, which previously could not all be read at once.

+++ src/SparkFunBMP384.cpp
@@ -375,7 +375,7 @@ uint8_t BMP384::readRegister(uint8_t regAddress)
     return data;
 }
 
-void BMP384::readRegisters(uint8_t regAddress, void* dataBuffer, uint8_t numBytes)
+void BMP384::readRegisters(uint8_t regAddress, void* dataBuffer, uint16_t numBytes)
 {
     switch(communicationInterface)
     {
